chore: Add items and points controllers
