+++ backend/src/routes.ts
@@ -1,60 +1,14 @@
 import express from 'express';
-import knex from './database/connection';
+
+import PointsController from './controllers/PointsController';
+import ItemsController from './controllers/ItemsControllers';
 
 const routes = express.Router();
+const pointsController = new PointsController();
+const itemsController = new ItemsController();
 
-routes.get('/items', async (request, response) => {
-  const items = await knex('items').select('*');
+routes.get('/items', itemsController.index);
 
-  const serializedItems = items.map(item => {
-    return {
-      id: item.id,
-      title: item.name,
-      image_url: `http://localhost:3333/uploads/${item.image}`,
-    }
-  })
-
-  response.json(serializedItems);
-});
-
-routes.post('/points', async (request, response) => {
-  const {
-    name,
-    email,
-    whatsapp,
-    latitude,
-    longitude,
-    city,
-    uf,
-    items
-  } = request.body;
-
-  const trx = await knex.transaction();
-
-  const insertedIds = await trx('points').insert({
-    image: 'image-fake',
-    name,
-    email,
-    whatsapp,
-    latitude,
-    longitude,
-    city,
-    uf
-  });
-
-  const point_id = insertedIds[0];
-
-  const pointItems = items.map((item_id: number) => {
-    return {
-      item_id,
-      point_id,
-    };
-  });
-
-  await trx('points_items').insert(pointItems);
-
-  return response.json({ success: true });
-
-});
+routes.post('/points', pointsController.create);
 
 export default routes;
